requirements.txt
fastapi
uvicorn
python-multipart
jinja2
gtts
openai-whisper
ffmpeg-python
google-generativeai
livekit-agents
silero-vad
torch
